@@ -2,7 +2,11 @@ import 'package:flutter/material.dart';
 
 import 'breakpoint.dart';
 import 'breakpoints.dart';
-import 'nav_destination.dart';
+
+enum AdaptiveScaffoldFeature {
+  navigationRail,
+  bottomNavigationBar,
+}
 
 class AdaptiveScaffold extends StatefulWidget {
   const AdaptiveScaffold({
@@ -23,25 +27,31 @@ class AdaptiveScaffold extends StatefulWidget {
   final Widget body;
 
   @override
-  _AdaptiveScaffoldState createState() => _AdaptiveScaffoldState();
+  State<AdaptiveScaffold> createState() => _AdaptiveScaffoldState();
+
+  static Set<AdaptiveScaffoldFeature> featuresFor(BuildContext context) {
+    final _InheritedAdaptiveFeatures? inherited = context.dependOnInheritedWidgetOfExactType<_InheritedAdaptiveFeatures>();
+    return inherited?.features ?? {};
+  }
 }
 
 class _AdaptiveScaffoldState extends State<AdaptiveScaffold> with SingleTickerProviderStateMixin {
 
   Widget _buildWithNav(BuildContext context) {
-    late bool showNavRail;
+    bool showNavRail;
     final Breakpoint breakpoint = BreakpointLayout.breakpointFor(context);
     switch (breakpoint) {
-      case DesktopSmallBreakpoint():
-        showNavRail = false;
-        break;
       case DesktopLargeBreakpoint():
         showNavRail = true;
         break;
+      case DesktopSmallBreakpoint():
+      default:
+        showNavRail = false;
+        break;
     }
 
-    final double railWidth = 72;
-    final double bottomNavHeight = kBottomNavigationBarHeight;
+    const double railWidth = 72;
+    const double bottomNavHeight = kBottomNavigationBarHeight;
     final int selectedIndex = widget.destinations.indexOf(widget.selectedDestination);
 
     return Scaffold(
@@ -58,7 +68,7 @@ class _AdaptiveScaffoldState extends State<AdaptiveScaffold> with SingleTickerPr
                 minWidth: railWidth,
                 maxWidth: railWidth,
                 child: NavigationRail(
-                  destinations: widget.destinations.map(NavigationDestination.toRailDestination).toList(),
+                  destinations: widget.destinations.map(_toRailDestination).toList(),
                   selectedIndex: selectedIndex,
                   onDestinationSelected: (int index) {
                     widget.onDestinationSelected(widget.destinations[index]);
@@ -67,7 +77,10 @@ class _AdaptiveScaffoldState extends State<AdaptiveScaffold> with SingleTickerPr
               ),
             ),
           ),
-          Expanded(child: widget.body),
+          _InheritedAdaptiveFeatures(
+            features: { showNavRail ? AdaptiveScaffoldFeature.navigationRail : AdaptiveScaffoldFeature.bottomNavigationBar },
+            child: Expanded(child: widget.body),
+          ),
         ],
       ),
       bottomNavigationBar: AnimatedContainer(
@@ -80,7 +93,7 @@ class _AdaptiveScaffoldState extends State<AdaptiveScaffold> with SingleTickerPr
             minHeight: bottomNavHeight,
             maxHeight: bottomNavHeight,
             child: BottomNavigationBar(
-              items: widget.destinations.map(NavigationDestination.toBottomNavItem).toList(),
+              items: widget.destinations.map(_toBottomNavItem).toList(),
               currentIndex: selectedIndex,
               showUnselectedLabels: true,
               onTap: (int index) {
@@ -108,4 +121,33 @@ class _AdaptiveScaffoldState extends State<AdaptiveScaffold> with SingleTickerPr
       ? _buildWithNav(context)
       : _buildWithoutNav(context);
   }
+}
+
+class _InheritedAdaptiveFeatures extends InheritedWidget {
+  const _InheritedAdaptiveFeatures({
+    Key? key,
+    required this.features,
+    required Widget child,
+  }) : super(key: key, child: child);
+
+  final Set<AdaptiveScaffoldFeature> features;
+
+  @override
+  bool updateShouldNotify(_InheritedAdaptiveFeatures oldWidget) {
+    return features != oldWidget.features;
+  }
+}
+
+NavigationRailDestination _toRailDestination(NavigationDestination destination) {
+  return NavigationRailDestination(
+    label: Text(destination.label),
+    icon: destination.icon,
+  );
+}
+
+BottomNavigationBarItem _toBottomNavItem(NavigationDestination destination) {
+  return BottomNavigationBarItem(
+    label: destination.label,
+    icon: destination.icon,
+  );
 }
